Почистил main.js
